dry up event waiting and http clients
